refactor: Rename assembly and namespace (MA -> AV3)
