+++ Packages/jp.suzuryg.face-emo/Editor/Detail/AV3/AvatarProcessor.cs
@@ -1,4 +1,4 @@
-﻿using Suzuryg.FaceEmo.Components.MA;
+﻿using Suzuryg.FaceEmo.Components.AV3;
 using System;
 using UnityEngine;
 using VRC.SDKBase.Editor.BuildPipeline;
