+++ Packages/jp.suzuryg.face-emo/Editor/Detail/AV3/FxGenerator.cs
@@ -2,8 +2,8 @@
 using AnimatorAsCode.V0.Extensions.VRChat;
 using Suzuryg.FaceEmo.Domain;
 using Suzuryg.FaceEmo.UseCase;
+using Suzuryg.FaceEmo.Components.AV3;
 using Suzuryg.FaceEmo.Components.Settings;
-using Suzuryg.FaceEmo.Components.MA;
 using Suzuryg.FaceEmo.Detail.Drawing;
 using Suzuryg.FaceEmo.External.Hai.ComboGestureIntegrator;
 using System;
