+++ Packages/jp.suzuryg.face-emo/Runtime/Components/AV3/RunBeforeModularAvatar.cs
@@ -5,7 +5,7 @@ using VRC.SDK3.Avatars.Components;
 using nadena.dev.modular_avatar.core;
 #endif
 
-namespace Suzuryg.FaceEmo.Components.MA
+namespace Suzuryg.FaceEmo.Components.AV3
 {
     [DefaultExecutionOrder(-20000)] // run before modular avatar
     public abstract class RunBeforeModularAvatar : MonoBehaviour , VRC.SDKBase.IEditorOnly // Inherit IEditorOnly to register in VRChat whitelist
