+++ Packages/jp.suzuryg.face-emo/Runtime/Components/AV3/TrackingControlDisabler.cs
@@ -11,7 +11,7 @@ using UnityEditor;
 using UnityEditor.Animations;
 #endif
 
-namespace Suzuryg.FaceEmo.Components.MA
+namespace Suzuryg.FaceEmo.Components.AV3
 {
     [DisallowMultipleComponent]
     public class TrackingControlDisabler : RunBeforeModularAvatar
