+++ Packages/jp.suzuryg.face-emo/Runtime/Components/AV3/jp.suzuryg.face-emo.components.av3.Runtime.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "jp.suzuryg.face-emo.components.ma.Runtime",
+    "name": "jp.suzuryg.face-emo.components.av3.Runtime",
     "references": [
         "GUID:ad29c40ce7d8264468d6c7634e6a3a40"
     ],
